pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docvector"
version = "0.1.0"
description = "Self-hostable documentation vector search system"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "DocVector Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.12.0",

    # Vector database
    "qdrant-client>=1.7.0",

    # Embeddings
    "sentence-transformers>=2.2.0",
    "httpx>=0.25.0",

    # Caching
    "redis[hiredis]>=5.0.0",

    # Web scraping
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "html2text>=2020.1.16",
    "crawl4ai>=0.4.0",

    # Text processing
    "markdown>=3.5.0",
    "markdown-it-py>=3.0.0",

    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",

    # MCP (Model Context Protocol)
    "mcp>=1.0.0",
]

[project.scripts]
docvector = "docvector.cli:app"
docvector-mcp = "docvector.mcp.server:main"

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "aioresponses>=0.7.6",

    # Code quality
    "black>=23.11.0",
    "ruff>=0.1.5",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
docvector = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=docvector",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src/docvector"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
